chore: organize csv into folder
