+++ src/std/index.ts
@@ -1,1 +1,1 @@
-export * as csv from './csv.js';
+export * as csv from './csv/index.js';
